// src/app/widgets/constants/widget-types.ts
export enum WidgetType {
  LiquidFillGauge = 'LiquidFillGauge',
  Bar = 'Bar',
  Billboard = 'Billboard',
  GeoMap = 'Geo Map',
  Line = 'Line',
  Tabular = 'Tabular',
  Sankey = 'Sankey',
  SolidGauge = 'Solid Gauge',
  Sunburst = 'Sunburst',
  TrendDiff = 'TrendDiff',
  CallTimeLine = 'Call TimeLine',
  LabelWidget = 'Label',
  Bubble = 'Bubble',
  EventViewer = 'Raw Event Viewer'
}

export enum WidgetMode {
  Measures = 'Measures',
  Instances = 'Instances',
  TimeRange = 'Time Range'
}

export enum WidgetTypeForGroup {
  Bar = 'Bar',
  Line = 'Line',
  GeoMap = 'Map',
  Tabular = 'Tabular',
  Sankey = 'Flow',
  SolidGauge = 'Gauge',
  Sunburst = 'Proportion',
  TrendDiff = 'Ranking',
  Bubble = 'Correlation',
}